+++ lib/mqtt-lib.js
@@ -66,6 +66,14 @@ ClientController.prototype.subscribe = function () {
   this.client.subscribe(this.topicName);
 };
 
+ClientController.prototype.subscribeByTopic = function (topicName) {
+  this.client.subscribe(topicName);
+};
+
+ClientController.prototype.publishByTopic = function (topicName, message) {
+  this.client.publish(topicName, message);
+};
+
 ClientController.prototype.publish = function (message) {
   this.client.publish(this.topicName, message);
 };
